refactor: unwrap me.universi.group.entities.GroupSettings package
