+++ src/main/java/me/universi/group/repositories/GroupEmailFilterRepository.java
@@ -1,7 +1,7 @@
 package me.universi.group.repositories;
 
 
-import me.universi.group.entities.GroupSettings.GroupEmailFilter;
+import me.universi.group.entities.GroupEmailFilter;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
