+++ src/main/java/me/universi/group/repositories/GroupEnvironmentRepository.java
@@ -1,6 +1,6 @@
 package me.universi.group.repositories;
 
-import me.universi.group.entities.GroupSettings.GroupEnvironment;
+import me.universi.group.entities.GroupEnvironment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
