+++ src/main/java/me/universi/group/repositories/GroupThemeRepository.java
@@ -1,7 +1,7 @@
 package me.universi.group.repositories;
 
 
-import me.universi.group.entities.GroupSettings.GroupTheme;
+import me.universi.group.entities.GroupTheme;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
