+++ src/main/java/me/universi/group/services/GroupEnvironmentService.java
@@ -3,8 +3,8 @@ package me.universi.group.services;
 import me.universi.api.exceptions.*;
 import me.universi.group.DTO.UpdateGroupEnvironmentDTO;
 import me.universi.group.entities.Group;
-import me.universi.group.entities.GroupSettings.GroupEnvironment;
-import me.universi.group.entities.GroupSettings.GroupSettings;
+import me.universi.group.entities.GroupEnvironment;
+import me.universi.group.entities.GroupSettings;
 import me.universi.group.repositories.GroupEnvironmentRepository;
 import me.universi.role.services.RoleService;
 import me.universi.user.services.UserService;
